membuat class baru yaitu personal
