+++ src/lib/Employee.java
@@ -7,22 +7,13 @@ import java.util.List;
 
 public class Employee {
 
-	private String employeeId;
-	private String firstName;
-	private String lastName;
-	private String idNumber;
-	private String address;
+
 	
 
 	private LocalDate joinedDate;
 	private int monthWorkingInYear;
 	
-	private boolean isForeigner;
-	public enum gender{
-		PRIA,
-		WANITA
-	}
-	public gender Gender;
+	private Personal personal
 	
 	private int monthlySalary;
 	private int otherMonthlyIncome;
@@ -33,6 +24,8 @@ public class Employee {
 
 	private List<String> childNames;
 	private List<String> childIdNumbers;
+	
+	
 	
 	public Employee(String employeeId, String firstName, String lastName, String idNumber, String address, int yearJoined, int monthJoined, int dayJoined, boolean isForeigner, boolean gender) {
 		this.employeeId = employeeId;
